Revert "Implements GetStyles request on WMS 1.1.1 [GEOT-5309]"
