+++ modules/extension/wms/src/test/java/org/geotools/data/wms/test/LocalGeoServerOnlineTest.java
@@ -19,16 +19,12 @@ package org.geotools.data.wms.test;
 import java.awt.Color;
 import java.awt.image.BufferedImage;
 import java.io.InputStream;
-import java.io.InputStreamReader;
 import java.net.MalformedURLException;
 import java.net.URL;
 import java.util.Collections;
-import java.util.HashMap;
-import java.util.Map;
 import java.util.Properties;
 
 import javax.imageio.ImageIO;
-import javax.mail.internet.ContentType;
 
 import junit.framework.TestCase;
 
@@ -36,15 +32,16 @@ import org.apache.commons.io.IOUtils;
 import org.apache.commons.io.output.StringBuilderWriter;
 import org.geotools.data.ResourceInfo;
 import org.geotools.data.ServiceInfo;
-import org.geotools.data.ows.*;
-import org.geotools.data.wms.WMS1_1_1;
+import org.geotools.data.ows.CRSEnvelope;
+import org.geotools.data.ows.Layer;
+import org.geotools.data.ows.MultithreadedHttpClient;
+import org.geotools.data.ows.OperationType;
+import org.geotools.data.ows.WMSCapabilities;
 import org.geotools.data.wms.WebMapServer;
 import org.geotools.data.wms.request.GetFeatureInfoRequest;
 import org.geotools.data.wms.request.GetMapRequest;
-import org.geotools.data.wms.request.GetStylesRequest;
 import org.geotools.data.wms.response.GetFeatureInfoResponse;
 import org.geotools.data.wms.response.GetMapResponse;
-import org.geotools.data.wms.response.GetStylesResponse;
 import org.geotools.factory.GeoTools;
 import org.geotools.factory.Hints;
 import org.geotools.geometry.GeneralEnvelope;
@@ -53,9 +50,6 @@ import org.geotools.referencing.CRS;
 import org.geotools.referencing.CRS.AxisOrder;
 import org.geotools.referencing.crs.DefaultEngineeringCRS;
 import org.geotools.referencing.crs.DefaultGeographicCRS;
-import org.geotools.styling.*;
-import org.geotools.styling.builder.*;
-import org.junit.Test;
 import org.opengis.geometry.Envelope;
 import org.opengis.referencing.FactoryException;
 import org.opengis.referencing.NoSuchAuthorityCodeException;
@@ -113,7 +107,6 @@ import org.opengis.util.GenericName;
  */
 public class LocalGeoServerOnlineTest extends TestCase {
     static private String LOCAL_GEOSERVER = "http://127.0.0.1:8080/geoserver/ows?SERVICE=WMS&";
-    static private String LOCAL_LAYERS = "test_shp:TRONCON_ROUTE";
 
     static private WebMapServer wms;
 
@@ -424,80 +417,4 @@ public class LocalGeoServerOnlineTest extends TestCase {
             fail( context+": GetFeatureInfo BBOX=" + envelope );
         }
     }
-
-    public void testGetStyle() throws Exception {
-
-        String baseUrl = LocalGeoServerOnlineTest.LOCAL_GEOSERVER;
-        String layers = LocalGeoServerOnlineTest.LOCAL_LAYERS;
-
-        URL url = new URL(baseUrl);
-
-        GetStylesResponse wmsResponse = null;
-        GetStylesRequest wmsRequest = null;
-        StyleFactory styleFactory = new StyleFactoryImpl();
-
-        WebMapServer server = new WebMapServer(url) {
-            // GetStyle is only implemented in WMS 1.1.1
-            protected void setupSpecifications() {
-                specs = new Specification[1];
-                specs[0] = new WMS1_1_1();
-            }
-        };
-
-        wmsRequest = server.createGetStylesRequest();
-        wmsRequest.setLayers(layers);
-        // Test URL
-        String queryParamters = wmsRequest.getFinalURL().getQuery();
-        Map parameters = new HashMap();
-        String[] rawParameters = queryParamters.split("&");
-        for(String param : rawParameters){
-            String [] keyValue = param.split("=");
-            parameters.put(keyValue[0],keyValue[1]);
-        }
-
-        assertEquals(4, parameters.size());
-        assertEquals("WMS", parameters.get("SERVICE"));
-        assertEquals("GetStyles", parameters.get("REQUEST"));
-        assertEquals("1.1.1", parameters.get("VERSION"));
-        assertEquals(layers, parameters.get("LAYERS"));
-
-        wmsResponse = server.issueRequest(wmsRequest);
-
-        // Set encoding of response from HTTP content-type header
-        ContentType contentType = new ContentType(wmsResponse.getContentType());
-        InputStreamReader stream;
-        if(contentType.getParameter("charset") != null)
-            stream = new InputStreamReader(wmsResponse.getInputStream(), contentType.getParameter("charset"));
-        else
-            stream = new InputStreamReader(wmsResponse.getInputStream());
-
-        Style[] styles = (new SLDParser(styleFactory, stream)).readXML();
-
-        assert styles.length > 0;
-
-        SLDTransformer styleTransform = new SLDTransformer();
-        StyledLayerDescriptorBuilder SLDBuilder = new StyledLayerDescriptorBuilder();
-
-        NamedLayerBuilder namedLayerBuilder = SLDBuilder.namedLayer();
-        namedLayerBuilder.name(layers);
-        org.geotools.styling.builder.StyleBuilder styleBuilder = namedLayerBuilder.style();
-
-        for(int i =0; i<styles.length; i++){
-            styleBuilder.reset(styles[i]);
-            styles[i] = styleBuilder.build();
-        }
-
-        NamedLayer namedLayer = namedLayerBuilder.build();
-
-        for(Style style: styles)
-            namedLayer.addStyle(style);
-
-        StyledLayerDescriptor sld = (new StyledLayerDescriptorBuilder()).build();
-        sld.addStyledLayer(namedLayer);
-        String xml = styleTransform.transform(sld);
-        assert xml.length() > 300;
-    }
-
-
-
 }
